i#1376 VS2013: build core with VS2013
For VS2013, do not link with MSVCR* and instead pull _aulldvrm from ntdll.
This precludes running on 2K or NT with a VS2013 build.  This requires
adding QIfist to utils.c, which necessitates using io.c's
double2int_trunc() in utils.c.

Use /FS for proper parallel builds with VS2013.

Fix misc warnings from the new compiler.

SVN-Revision: 2648

+++ core/x86/arch_exports.h
@@ -249,7 +249,7 @@ emit_detach_callback_final_jmp(dcontext_t *dcontext,
     ASSERT(sizeof(value) == 4);                                     \
     /* test that we aren't crossing a cache line boundary */        \
     CHECK_JMP_TARGET_ALIGNMENT(target, 4, hot_patch);               \
-    _InterlockedExchange((volatile LONG *)target, value);           \
+    _InterlockedExchange((volatile LONG *)target, (LONG)value);     \
   } while (0)
 # ifdef X64
 #  define ATOMIC_8BYTE_WRITE(target, value, hot_patch) do {         \
